refactor: 拆分 storage/todo 纯 helper 与行映射
将 storage/todo/mod.rs 中的排序、ID 解析、时间显示、草稿规范化、搜索打分、
SQL 查询与行映射抽到 id/normalize/query/search/sort/time 子模块，TodoStore
对外 API 与 schema 保持不变，相关持久化、脱敏、状态转换注释一并迁移。

refs #115
